requirements.txt
newspaper3k==0.2.6
nltk==3.3
mysql-connector==2.1.6
openpyxl==2.5.4
xlwt==1.3.0